Adicion de routes para metodos HTTP del microservicio de adopcion
